NAM-33
Added a few validation and check changes to dislay the correct quota percetage in the correct row. Also added and commented out the new code that will check if the original marker is still available and if the quota value needs to be added to the reassigned marker's value. Also tried adding the assignmentmarker object to the VM file, but that's still breaking.